# Shortcut Cast.toNumber if given a number