### Resolves

Close #1705 

### Proposed Changes

Check if `typeof value === 'number'` and do a separate isNaN check to confirm its not NaN.

### Reason for Changes

If toNumber is called on a number avoiding passing the number to Number() can provide a small performance improvement.

### Numbers

This change alone does not improve the blocks per second or fps in Scratch due to how the Scratch VM runs. Scratch iterates all threads before checking if it has used too much time instead of checking after each thread. I mention the later because that would be more likely to show a difference in blocks per second or fps (and we don't use that method because it would make projects nondeterministic).

With that in mind for this set of performance numbers I've recorded the percent of time spent in Cast.toNumber with browser inspector performance panels.

Together with other changes this will help speed up Scratch.

|  Project, Platform | before (% of time) | after (% of time) | Change |
|  ------ | ------ | ------ | ------ |
|  89811578, 2017 MBP 15" Chrome | 1.50% | 1.04% | 0.46% |
|  89811578, 2017 MBP 15" Firefox | 1.79% | 1.16% | 0.63% |
|  89811578, 2017 MBP 15" Safari | 2.00% | 2.00% | 0.00% |
|  89811578, 2017 SMS Chromebook | 1.58% | 0.90% | 0.68% |
|  89811578, 2018 Raspberry Pi 3B+ | 1.33% | 1.12% | 0.21% |